sound/soc: aw882xx: Don't use max amplifier gain for AW882XX_RECEIVER_MODE
In the case of Fairphone 4's stereo speaker setup, where one speaker
also serves as the earpiece module, the earpiece is treated as a mono
transmit source during standalone use, such as in VoIP calls. The
current issue lies in the amplifier applying maximum gains regardless of
whether it operates standalone or as part of a stereo setup which is not
ideal considering the usecase.

Controlling amplifier gains for a specific audio backend device via
mixer paths is not possible, as the property exported by the amplifier
gets overwritten by the kernelspace.

When the "handset" device is being used the amplifier switches the
working mode to "Rcv" as shown in [1], this patch checks whether the
current working mode is "Rcv" and limits the gain being set to max.

The more appropriate fix is adjusting digital RX volume or addressing
and tuning the volume curves for the HANDSET backend on ACDBs.

[1] https://github.com/WeAreFairphone/android_device_fairphone_FP4/blob/staging/lineage-19.1/audio/mixer_paths_lagoon_fp4.xml#L3428